auth+lsat: move LSAT specific code
