+++ lsat/context.go
@@ -1,4 +1,4 @@
-package auth
+package lsat
 
 import (
 	"context"
