+++ proxy/proxy.go
@@ -12,6 +12,7 @@ import (
 	"strings"
 
 	"github.com/lightninglabs/aperture/auth"
+	"github.com/lightninglabs/aperture/lsat"
 	"google.golang.org/grpc/codes"
 )
 
@@ -184,12 +185,12 @@ func (p *Proxy) director(req *http.Request) {
 
 		// Make sure we always forward the authorization in the correct/
 		// default format so the backend knows what to do with it.
-		mac, preimage, err := auth.FromHeader(&req.Header)
+		mac, preimage, err := lsat.FromHeader(&req.Header)
 		if err == nil {
 			// It could be that there is no auth information because
 			// none is needed for this particular request. So we
 			// only continue if no error is set.
-			err := auth.SetHeader(&req.Header, mac, preimage)
+			err := lsat.SetHeader(&req.Header, mac, preimage)
 			if err != nil {
 				log.Errorf("could not set header: %v", err)
 			}
